backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.20
pydantic-settings==2.9.1
numpy==2.0.2
httpx==0.28.1
